+++ src/services/index.ts
@@ -1,3 +1,4 @@
+export * from "./BackgroundService";
 export * from "./CommandsService";
 export * from "./ConfigListeners";
 export * from "./DocHubTreeviewProvider";